+++ inst/tinytest/test_ctrdata_ctrfindactivesubstance.R
@@ -26,3 +26,9 @@ expect_error(
   ctrFindActiveSubstanceSynonyms(
     activesubstance = iris)
 )
+
+# test
+expect_null(
+  suppressMessages(
+    ctrFindActiveSubstanceSynonyms("ISURELYDONOTEXIST"))
+)