+++ NEWS.md
@@ -1,5 +1,6 @@
 # ctrdata 1.9.1.9000
  - 2022-04-24
+ - ctrFindActiveSubstanceSynonyms() returns NULL for non-existing active substance
  
 # ctrdata 1.9.1
  - 2022-04-24